mixed refactoring
