+++ .gitignore
@@ -2,6 +2,7 @@
 .DS_Store
 .idea
 *.db
+.vscode
 
 # Byte-compiled / optimized / DLL files
 __pycache__/
