+++ README.md
@@ -1,1 +1,6 @@
 # AnpaBot
+
+Open Source bot for Atlantic Gaming
+
+
+
