+++ anpabot/cogs/botadmin.py
@@ -1,9 +1,10 @@
 
 from discord.ext import commands
 from discord.ext.commands import Bot
-from anpabot.configstore import ConfigStore
+from anpabot.persistence.configstore import ConfigStore
 
 
+"""Module to manage the anpabot"""
 class BotAdmin(commands.Cog):
     def __init__(self, bot: Bot, config: ConfigStore):
         self.bot = bot
@@ -14,24 +15,6 @@ class BotAdmin(commands.Cog):
         """Add role to control the bot, use like: `addadmin admin_role`"""
         if self._config.is_admin(ctx.author):
             message = self._config.add_admin(ctx.guild, role_name)
-            await ctx.channel.send(message)
-        else:
-            await ctx.channel.send(f'Sorry {ctx.author.display_name} you don\'t have permissions')
-
-    @commands.command()
-    async def setjoinrole(self, ctx, role_name: str):
-        """Set default role for new joiners, use like: `setjoinrole default_role`"""
-        if self._config.is_admin(ctx.author):
-            message = self._config.set_default_role(ctx.guild, role_name)
-            await ctx.channel.send(message)
-        else:
-            await ctx.channel.send(f'Sorry {ctx.author.display_name} you don\'t have permissions')
-
-    @commands.command()
-    async def setmemberrole(self, ctx, role_name: str):
-        """Set role for members after accepting the rules, use like: `setmemberrole member_role`"""
-        if self._config.is_admin(ctx.author):
-            message = self._config.set_member_role(ctx.guild, role_name)
             await ctx.channel.send(message)
         else:
             await ctx.channel.send(f'Sorry {ctx.author.display_name} you don\'t have permissions')
